+++ ICD.Connect.Cameras.Vaddio/Properties/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 using System.Reflection;
 
-[assembly: AssemblyTitle("ICD.Connect.Cameras.Visca")]
+[assembly: AssemblyTitle("ICD.Connect.Cameras.Vaddio")]
 [assembly: AssemblyCompany("ICD Systems")]
-[assembly: AssemblyProduct("ICD.Connect.Cameras.Visca")]
+[assembly: AssemblyProduct("ICD.Connect.Cameras.Vaddio")]
 [assembly: AssemblyCopyright("Copyright © ICD Systems 2018")]
 [assembly: AssemblyVersion("3.3.2.0")]