+++ CHANGELOG.md
@@ -6,6 +6,30 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.3.2] - 2018-11-08
+### Changed
+ - Fail gracefully when loading a device and the port is not available
+
+## [3.3.1] - 2018-10-04
+### Changed
+ - Vaddio Roboshot has default username, password and ptz speeds
+
+## [3.3.0] - 2018-09-25
+### Added
+ - Vaddio Roboshot driver
+
+## [3.2.2] - 2018-09-14
+### Changed
+ - Performance improvements for camera routing
+
+## [3.2.1] - 2018-07-02
+### Changed
+ - Various Visca camera driver fixes
+
+## [3.2.0] - 2018-06-04
+### Changed
+ - Serial devices use ConnectionStateManager for maintaining connection to remote endpoints
+ 
 ## [3.1.0] - 2018-05-09
 ### Changed
  - Camera preset events are re-raised from the device